% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Greet an entity. Could be a person, an animal, or an alien.}
\usage{
hello(name)
}
\arguments{
\item{name}{The name of the entity to greet.}
}
\value{
A greeting as a string.
}
\description{
Greet an entity. Could be a person, an animal, or an alien.
}
\examples{
hello("E.T.")
}
